+++ internal/compile/manager_test.go
@@ -336,7 +336,7 @@ func (ms *MockStore) AddOrUpdateSchema(ctx context.Context, schemas ...*schemav1
 	return nil
 }
 
-func (ms *MockStore) Disable(ctx context.Context, _ ...string) (int64, error) {
+func (ms *MockStore) Disable(ctx context.Context, _ ...string) (uint32, error) {
 	args := ms.MethodCalled("Disable", ctx)
 	if res := args.Get(0); res == nil {
 		return 0, args.Error(0)
